AI/to_be_deleted/AI_Module_test/requirements.txt
numpy==1.26.4
pandas==2.2.1
lightgbm==4.3.0
matplotlib==3.8.3
plotly==5.19.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
jupyter==1.0.0
notebook==7.1.1
scikit-learn==1.4.1.post1
httpx==0.27.*